Adding a new mimage (leg/copy) to a mirror behaves differently depending on if the mirror has a 'core' or 'disk' log.  When there is a disk log, the new leg is added by stacking a new mirror on top of the old (one leg is the old mirror and the other leg is the newly added device).  When the log is a 'core' log, the new leg is simply added to the existing mirror and all the devices are re-synced.
The logic that handles collapsing the stacked 'disk' log mirror was
having the effect of causing 'core' logged mirrors to begin resync'ing
for a second time.  I have used the 'CONVERTING' flag to indicate that
a mirror is converting by way of stacking.  This is no longer set for
up-converting core logs.  The final 'collapse' logic can safely be skipped
for 'core' log mirrors - getting rid of the second resync.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>